+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Changes from PR
+### Added
+- Logging errors instead of responsemessage when ingest throws an error
+
 ## [2.0.0 - 2022-10-21]
 ### Added
 - Added support for Nested Content (Umbraco 7)